ENH: updated CUDA Driver API to version 3020 and added support for _v2 suffixed Driver API functions
Former-commit-id: f5cf5a4dd299f730099b73348239d3802fbe6148

+++ ocelot/ocelot/cuda/interface/CudaDriverInterface.h
@@ -46,7 +46,7 @@ namespace cuda
 			CUresult cuDeviceGetName(char *name, int len, CUdevice dev);
 			CUresult cuDeviceComputeCapability(int *major, int *minor, 
 				CUdevice dev);
-			CUresult cuDeviceTotalMem(unsigned int *bytes, CUdevice dev);
+			CUresult cuDeviceTotalMem(size_t *bytes, CUdevice dev);
 			CUresult cuDeviceGetProperties(CUdevprop *prop, 
 				CUdevice dev);
 			CUresult cuDeviceGetAttribute(int *pi, 
@@ -86,7 +86,7 @@ namespace cuda
 			CUresult cuModuleGetFunction(CUfunction *hfunc, 
 				CUmodule hmod, const char *name);
 			CUresult cuModuleGetGlobal(CUdeviceptr *dptr, 
-				unsigned int *bytes, CUmodule hmod, const char *name);
+				size_t *bytes, CUmodule hmod, const char *name);
 			CUresult cuModuleGetTexRef(CUtexref *pTexRef, CUmodule hmod, 
 				const char *name);
 
@@ -96,20 +96,20 @@ namespace cuda
 			**
 			***********************************/
 
-			CUresult cuMemGetInfo(unsigned int *free, 
-				unsigned int *total);
+			CUresult cuMemGetInfo(size_t *free, 
+				size_t *total);
 
 			CUresult cuMemAlloc( CUdeviceptr *dptr, 
 				unsigned int bytesize);
 			CUresult cuMemAllocPitch( CUdeviceptr *dptr, 
-						          unsigned int *pPitch,
+						          size_t *pPitch,
 						          unsigned int WidthInBytes, 
 						          unsigned int Height, 
 						          unsigned int ElementSizeBytes
 						         );
 			CUresult cuMemFree(CUdeviceptr dptr);
 			CUresult cuMemGetAddressRange( CUdeviceptr *pbase, 
-				unsigned int *psize, CUdeviceptr dptr );
+				size_t *psize, CUdeviceptr dptr );
 
 			CUresult cuMemAllocHost(void **pp, unsigned int bytesize);
 			CUresult cuMemFreeHost(void *p);
@@ -272,7 +272,7 @@ namespace cuda
 
 			CUresult cuTexRefSetArray( CUtexref hTexRef, CUarray hArray, 
 				unsigned int Flags );
-			CUresult cuTexRefSetAddress( unsigned int *ByteOffset, 
+			CUresult cuTexRefSetAddress( size_t *ByteOffset, 
 				CUtexref hTexRef, CUdeviceptr dptr, unsigned int bytes );
 			CUresult cuTexRefSetAddress2D( CUtexref hTexRef, 
 				const CUDA_ARRAY_DESCRIPTOR *desc, CUdeviceptr dptr, 
@@ -364,7 +364,7 @@ namespace cuda
 				CUarray *pArray, CUgraphicsResource resource, 
 				unsigned int arrayIndex, unsigned int mipLevel );
 			CUresult cuGraphicsResourceGetMappedPointer(
-				CUdeviceptr *pDevPtr, unsigned int *pSize, 
+				CUdeviceptr *pDevPtr, size_t *pSize, 
 				CUgraphicsResource resource );
 			CUresult cuGraphicsResourceSetMapFlags(
 				CUgraphicsResource resource, unsigned int flags );
